# swagger.yaml
openapi: 3.0.0
info:
  title: Invoice backend
  description:  Invoice API docs
  version: 1.0.0
paths:
...
  /invoice:
    post:
      tags:
      - "invoice"
      description: "Create invoices"
      requestBody:
       description: Schema of invoice
       required: true
       content:
        application/json:
         schema:
          $ref: '#/components/schemas/Invoice'
      responses:
        "201":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 201
                 message:
                   type: "string"
                   example: request successful invoice created succesfully
               pdf: 
                type: "string"
               data:
                type: "object"
                properties:
                 _id:
                  type: "string"
                  example: "61efaca536255b2a0421b837"
                 client:
                  type: "string"
                  example: "treniq"
                 invoice_number:
                  type: "string"
                  example: "61efaca536255b2a0421b838"
                 bill_from:
                  type: "string"
                  example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                  Nagin street, Sehari, Guntur Dt, AP, 524560
                  8130657523
                  example@gmail.com
                  PAN: ADIBC9018B
                  GSTIN: 37AAIBB8121B1Z1"
                 bill_to:
                  type: "string"
                  example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                  Nagin street, Sehari, Guntur Dt, AP, 524560
                  8130657523
                  example@gmail.com
                  PAN: ADIBC9018B
                  GSTIN: 37AAIBB8121B1Z1"
                 ship_to:
                  type: "string"
                  example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                  Nagin street, Sehari, Guntur Dt, AP, 524560
                  8130657523
                  example@gmail.com
                  PAN: ADIBC9018B
                  GSTIN: 37AAIBB8121B1Z1"
                 payment_terms:
                  type: "string"
                  example: "payment terms"
                 date:
                  type: "string"
                  example: "1/7/2021"
                 due_date:
                  type: "string"
                  example: "1/7/2021"
                 items:
                  type: array
                  items:
                   type: object
                   properties:
                    item:
                     type: "string"
                     example: "item"
                    quantity:
                     type: "string"
                     example: "20"
                    rate:
                     type: "string"
                     example: "₹30"
                    amount:
                     type: "string"
                     example: "₹600"
                 sub_total:
                  type: "string"
                  example: "₹600"
                 tax:
                  type: "string"
                  example: "10%"
                 total:
                  type: "string"
                  example: "₹600"
                 discount:
                  type: "string"
                  example: "10%"
                 amount_paid:
                  type: "string"
                  example: "₹100"
                 balance_due:
                  type: "string"
                  example: "₹500"
                 notes:
                  type: "string"
                  example: "10%"
                 terms:
                  type: "string"
                  example: "terms" 
                 __v:
                  type: "number"
                  example: 0
                
        "400":
          description: Bad request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: Please enter ____
      
                        
    get:
      tags:
      - "invoice"
      description: "Get All invoices"
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful 
               data:
                type: "object"
                properties:
                 results:
                  type: "array"
                  items:
                   type: "object"
                   properties:
                    _id:
                     type: "string"
                     example: "61efaca536255b2a0421b837"
                    client:
                     type: "string"
                     example: "treniq"
                    invoice_number:
                     type: "string"
                     example: "61efaca536255b2a0421b838"
                    bill_from:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    bill_to:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    ship_to:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    payment_terms:
                     type: "string"
                     example: "payment terms"
                    date:
                     type: "string"
                     example: "1/7/2021"
                    due_date:
                     type: "string"
                     example: "1/7/2021"
                    items:
                     type: array
                     items:
                      type: object
                      properties:
                       item:
                        type: "string"
                        example: "item"
                       quantity:
                        type: "string"
                        example: "20"
                       rate:
                        type: "string"
                        example: "₹30"
                       amount:
                        type: "string"
                        example: "₹600"
                    sub_total:
                     type: "string"
                     example: "₹600"
                    tax:
                     type: "string"
                     example: "10%"
                    total:
                     type: "string"
                     example: "₹600"
                    discount:
                     type: "string"
                     example: "10%"
                    amount_paid:
                     type: "string"
                     example: "₹100"
                    balance_due:
                     type: "string"
                     example: "₹500"
                    notes:
                     type: "string"
                     example: "10%"
                    terms:
                     type: "string"
                     example: "terms"        
                
        "400":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: not found
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
  /invoice/{invoice_id}:
    get:
      tags:
      - "invoice"
      description: "Get invoice by Id"
      parameters:
      - name: invoice_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful
               data:
                type: "object"
                properties:
                    _id:
                     type: "string"
                     example: "61efaca536255b2a0421b837"
                    client:
                     type: "string"
                     example: "treniq"
                    invoice_number:
                     type: "string"
                     example: "61efaca536255b2a0421b838"
                    bill_from:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    bill_to:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    ship_to:
                     type: "string"
                     example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
                     Nagin street, Sehari, Guntur Dt, AP, 524560
                     8130657523
                     example@gmail.com
                     PAN: ADIBC9018B
                     GSTIN: 37AAIBB8121B1Z1"
                    payment_terms:
                     type: "string"
                     example: "payment terms"
                    date:
                     type: "string"
                     example: "1/7/2021"
                    due_date:
                     type: "string"
                     example: "1/7/2021"
                    items:
                     type: array
                     items:
                      type: object
                      properties:
                       item:
                        type: "string"
                        example: "item"
                       quantity:
                        type: "string"
                        example: "20"
                       rate:
                        type: "string"
                        example: "₹30"
                    amount:
                     type: "string"
                     example: "₹600"
                    sub_total:
                     type: "string"
                     example: "₹600"
                    tax:
                     type: "string"
                     example: "10%"
                    total:
                     type: "string"
                     example: "₹600"
                    discount:
                     type: "string"
                     example: "10%"
                    amount_paid:
                     type: "string"
                     example: "₹100"
                    balance_due:
                     type: "string"
                     example: "₹500"
                    notes:
                     type: "string"
                     example: "10%"
                    terms:
                     type: "string"
                     example: "terms"  
                
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: invoice not found         
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message  
          
    delete:
      tags:
      - "invoice"
      description: "Delete Invoice"
      parameters:
      - name: invoice_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 204
                 message:
                   type: "string"
                   example: request successful invoice deleted successfully
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: invoice not found  
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
...
components:
  schemas:
...
    Invoice:
     type: "object"
     required:
     - "client"
     - "invoice_number"
     - "bill_from"
     - "bill_to"
     - "payment_terms"
     - "date"
     - "due_date"
     - "items"
     - "sub_total"
     - "total"
     - "amount_paid"
     - "balance_due"
     properties:
      client:
        type: "string"
        example: "treniq"
      invoice_number:
        type: "string"
        example: "61efaca536255b2a0421b838"
      bill_from:
        type: "string"
        example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
        Nagin street, Sehari, Guntur Dt, AP, 524560
        8130657523
        example@gmail.com
        PAN: ADIBC9018B
        GSTIN: 37AAIBB8121B1Z1"
      bill_to:
        type: "string"
        example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
        Nagin street, Sehari, Guntur Dt, AP, 524560
        8130657523
        example@gmail.com
        PAN: ADIBC9018B
        GSTIN: 37AAIBB8121B1Z1"
      ship_to:
        type: "string"
        example: "CODEUNITY TECHNOLOGIES PVT LTD 1/42,
        Nagin street, Sehari, Guntur Dt, AP, 524560
        8130657523
        example@gmail.com
        PAN: ADIBC9018B
        GSTIN: 37AAIBB8121B1Z1"
      payment_terms:
        type: "string"
        example: "payment terms"
      date:
        type: "string"
        example: "1/7/2021"
      due_date:
        type: "string"
        example: "1/7/2021"
      items:
        type: array
        items:
         type: object
         properties:
          item:
           type: "string"
           example: "item"
          quantity:
           type: "string"
           example: "20"
          rate:
           type: "string"
           example: "₹30"
          amount:
           type: "string"
           example: "₹600"
      sub_total:
           type: "string"
           example: "₹600"
      tax:
           type: "string"
           example: "10%"
      total:
           type: "string"
           example: "₹600"
      discount:
           type: "string"
           example: "10%"
      amount_paid:
           type: "string"
           example: "₹100"
      balance_due:
           type: "string"
           example: "₹500"
      notes:
           type: "string"
           example: "10%"
      terms:
           type: "string"
           example: "terms"